changed navigation logic
