+++ plop-templates/component/plop.html
@@ -1,1 +1,1 @@
-<div>Template Component</div>
+<div>Template Component {{properCase name}}</div>
