+++ src/pages/dashboard/dashboard.html
@@ -1,2 +1,2 @@
-<ion-header>Template Page</ion-header>
-<ion-content></ion-content>
+<ion-header></ion-header>
+<ion-content>Template Page Dashboard</ion-content>
