+++ src/pages/sign-in/sign-in.html
@@ -10,6 +10,5 @@
 
   <ion-input type=email placeholder="Email" [(ngModel)]="email"></ion-input>
   <ion-input type=password placeholder="Password" [(ngModel)]="password"></ion-input>
-  <p (click)="logout()">Logout</p>
   <button ion-button full large class="full-bottom-btn" (click)="login()">Login</button>
 </ion-content>
